+++ packages/docsearch-react/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './DocSearchHit';
+export * from './DocSearchState';
 export * from './InternalDocSearchHit';
 export * from './StoredDocSearchHit';