+++ packages/docsearch-react/src/utils/groupBy.ts
@@ -1,6 +1,7 @@
 export function groupBy<TValue extends Record<string, unknown>>(
   values: TValue[],
-  predicate: (value: TValue) => string
+  predicate: (value: TValue) => string,
+  maxResultsPerGroup?: number
 ): Record<string, TValue[]> {
   return values.reduce<Record<string, TValue[]>>((acc, item) => {
     const key = predicate(item);
@@ -11,7 +12,7 @@ export function groupBy<TValue extends Record<string, unknown>>(
 
     // We limit each section to show 5 hits maximum.
     // This acts as a frontend alternative to `distinct`.
-    if (acc[key].length < 5) {
+    if (acc[key].length < (maxResultsPerGroup || 5)) {
       acc[key].push(item);
     }